+++ app/components/portfolios/positions_component.rb
@@ -12,9 +12,10 @@ module Portfolios
 
     def prepare_portfolio_stats
       @portfolio_stats = @positions.group_by(&:quote).each_with_object(basis_stats) do |(quote, positions), acc|
+        positions_stats = calculate_positions_stats(positions)
         acc[quote.security.type][:elements][quote] = {
-          total_price_cents: positions.sum(&:price_cents),
-          total_amount:      positions.sum(&:amount),
+          total_price_cents: positions_stats[:total_price_cents],
+          total_amount:      positions_stats[:total_amount],
           positions:         positions
         }
       end
@@ -25,6 +26,22 @@ module Portfolios
         'Share'      => { elements: {} },
         'Foundation' => { elements: {} },
         'Bond'       => { elements: {} }
+      }
+    end
+
+    def calculate_positions_stats(positions)
+      positions.each_with_object(basis_positions_stats) do |position, acc|
+        not_sold_amount = position.amount - position.sold_amount
+
+        acc[:total_amount]      += not_sold_amount
+        acc[:total_price_cents] += not_sold_amount * position.price_cents
+      end
+    end
+
+    def basis_positions_stats
+      {
+        total_price_cents: 0,
+        total_amount:      0
       }
     end
   end