+++ app/services/positions/create_service.rb
@@ -4,12 +4,19 @@ module Positions
   class CreateService
     prepend BasicService
 
-    def call(portfolio:, quote:, price:, amount:)
-      portfolio.positions.create(
-        quote:  quote,
-        price:  price,
-        amount: amount
-      )
+    def initialize(
+      buy_service:  Creation::BuyService,
+      sell_service: Creation::SellService
+    )
+      @buy_service  = buy_service
+      @sell_service = sell_service
+    end
+
+    def call(args={})
+      case args[:operation]
+      when '0' then @buy_service.call(args.except(:operation))
+      when '1' then @sell_service.call(args.except(:operation))
+      end
     end
   end
 end