+++ app/services/positions/creation/buy_service.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Positions
+  module Creation
+    class BuyService < BaseService
+      def call(args={})
+        create_position(args)
+      end
+    end
+  end
+end