+++ app/services/positions/fetching/for_portfolio_service.rb
@@ -6,7 +6,13 @@ module Positions
       prepend BasicService
 
       def call(user:)
-        @result = user.positions.order(id: :desc).includes(quote: :security)
+        @result =
+          user
+          .positions
+          .buying
+          .with_unsold_securities
+          .order(id: :desc)
+          .includes(quote: :security)
       end
     end
   end